tmpop, tmpoptestcases: moved tmpop tests into tmpoptestcases package
