+++ filestore/filestore_test.go
@@ -15,21 +15,22 @@
 package filestore
 
 import (
-	"os"
 	"testing"
 
-	"github.com/stratumn/sdk/store"
 	"github.com/stratumn/sdk/store/storetestcases"
+	"github.com/stratumn/sdk/tmpop/tmpoptestcases"
 )
 
 func TestFilestore(t *testing.T) {
 	storetestcases.Factory{
-		New: func() (store.Adapter, error) {
-			return createAdapter(t), nil
-		},
-		Free: func(s store.Adapter) {
-			a := s.(*FileStore)
-			defer os.RemoveAll(a.config.Path)
-		},
+		New:  createAdapter,
+		Free: freeAdapter,
+	}.RunTests(t)
+}
+
+func TestFileTMPop(t *testing.T) {
+	tmpoptestcases.Factory{
+		New:  createAdapter,
+		Free: freeAdapter,
 	}.RunTests(t)
 }
